requirements.txt
Flask==1.1.2
geographiclib==1.50
geopy==2.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
Werkzeug==1.0.1